+++ src/Thycotic.RabbitMq.Helper.PSCommands/Installation/GetRabbitMqInstallerCommand.cs
@@ -1,6 +1,9 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.Management.Automation;
 using System.Threading;
+using Thycotic.RabbitMq.Helper.Logic;
+using Thycotic.RabbitMq.Helper.Logic.IO;
 
 namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
 {
@@ -36,7 +39,6 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
         /// </summary>
         public static readonly string RabbitMqInstallerPath = Path.Combine(Path.GetTempPath(), string.Format("rabbitMq{0}.exe", InstallationConstants.RabbitMq.Version));
 
-
         /// <summary>
         ///     Gets or sets the offline rabbit mq installer path.
         /// </summary>
@@ -50,6 +52,12 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Offline)]
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("OfflinePath")]
         public string OfflineRabbitMqInstallerPath { get; set; }
 
@@ -64,6 +72,10 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Online)]
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("ForceDownload")]
         public SwitchParameter Force { get; set; }
 
@@ -79,8 +91,25 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
              ValueFromPipeline = true,
              ValueFromPipelineByPropertyName = true,
              ParameterSetName = ParameterSets.Online)]
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
         [Alias("Mirror")]
         public SwitchParameter UseThycoticMirror { get; set; }
+
+        /// <summary>
+        /// Gets or sets whether to prepare for offline install.
+        /// </summary>
+        /// <value>
+        ///   <c>true</c> if [force download]; otherwise, <c>false</c>.
+        /// </value>
+        /// <para type="description">Gets or sets whether to prepare for offline install.</para>
+        [Parameter(
+            ValueFromPipeline = true,
+            ValueFromPipelineByPropertyName = true,
+            ParameterSetName = ParameterSets.Prepare)]
+        public SwitchParameter PrepareForOfflineInstall { get; set; }
 
         /// <summary>
         ///     Processes the record.
@@ -90,16 +119,53 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
         {
             if (!string.IsNullOrWhiteSpace(OfflineRabbitMqInstallerPath))
             {
-                WriteVerbose(string.Format("Using offline installer path {0}", OfflineRabbitMqInstallerPath));
-
-                if (!File.Exists(OfflineRabbitMqInstallerPath))
+                if (PrepareForOfflineInstall)
                 {
-                    throw new FileNotFoundException("Installer does not exist");
-                }
-                if (File.Exists(RabbitMqInstallerPath))
-                    File.Delete(RabbitMqInstallerPath);
+                    WriteDebug($"Preparing offline installer path {OfflineRabbitMqInstallerPath}");
 
-                File.Copy(OfflineRabbitMqInstallerPath, RabbitMqInstallerPath);
+                    if (Force)
+                        WriteVerbose("Forcing download");
+
+                    WriteVerbose("Downloading Erlang");
+
+                    var downloader = new PrerequisiteDownloader();
+
+                    var downloadUrl = UseThycoticMirror
+                        ? InstallationConstants.RabbitMq.ThycoticMirrorDownloadUrl
+                        : InstallationConstants.RabbitMq.DownloadUrl;
+
+                    downloader.Download(CancellationToken.None, new Uri(downloadUrl, UriKind.Absolute),
+                        OfflineRabbitMqInstallerPath, InstallationConstants.RabbitMq.InstallerChecksum, Force, 5,
+                        WriteDebug, WriteVerbose, (s, exception) => throw exception,
+                        progress =>
+                        {
+                            WriteProgress(new ProgressRecord(1, "RabbitMq download in progress", "Downloading")
+                            {
+                                PercentComplete = progress.ProgressPercentage
+                            });
+                        });
+
+                }
+                else
+                {
+                    WriteVerbose(string.Format("Using offline installer path {0}", OfflineRabbitMqInstallerPath));
+
+                    if (!File.Exists(OfflineRabbitMqInstallerPath))
+                    {
+                        throw new FileNotFoundException("Installer does not exist");
+                    }
+
+                    if (PrerequisiteDownloader.CalculateMD5(OfflineRabbitMqInstallerPath) !=
+                        InstallationConstants.RabbitMq.InstallerChecksum)
+                    {
+                        throw new FileNotFoundException("Installer checksum does not match");
+                    }
+
+                    if (File.Exists(RabbitMqInstallerPath))
+                        File.Delete(RabbitMqInstallerPath);
+
+                    File.Copy(OfflineRabbitMqInstallerPath, RabbitMqInstallerPath);
+                }
             }
             else
             {
@@ -114,8 +180,8 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
                     ? InstallationConstants.RabbitMq.ThycoticMirrorDownloadUrl
                     : InstallationConstants.RabbitMq.DownloadUrl;
 
-                downloader.Download(CancellationToken.None, downloadUrl,
-                    RabbitMqInstallerPath, Force, 5, WriteDebug, WriteVerbose, (s, exception) => { throw exception; },
+                downloader.Download(CancellationToken.None, new Uri(downloadUrl, UriKind.Absolute),
+                    RabbitMqInstallerPath, InstallationConstants.RabbitMq.InstallerChecksum, Force, 5, WriteDebug, WriteVerbose, (s, exception) => throw exception,
                     progress =>
                     {
                         WriteProgress(new ProgressRecord(1, "RabbitMq download in progress", "Downloading")
@@ -123,6 +189,7 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
                             PercentComplete = progress.ProgressPercentage
                         });
                     });
+
             }
         }
 
@@ -130,6 +197,7 @@ namespace Thycotic.RabbitMq.Helper.PSCommands.Installation
         {
             public const string Offline = "Offline";
             public const string Online = "Online";
+            public const string Prepare = "Prepare";
         }
     }
 }
